third_party/blink/web_tests/svg/dynamic-updates/SVGMarkerElement-svgdom-refX-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'refX' property of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.refX.baseVal.value is 500
PASS markerElement.refX.baseVal.value is 5
PASS successfullyParsed is true

TEST COMPLETE
